Basic Foraging Interaction
Currently you can manually set up Foragables which can be interacted with. In the future it will hold an item but currently it holds a string.

The purpose of this is to test whether passing data between scenes is working, you can get rid of the item on level 1 and see if it stays gone