Equipment menu, pause menu improvements
The settings menu has a list of tabs that can be cycled through.

I added the equipment menu to the pause menu

The settings menu is overhauled to the pause menu now (it doesnt actually pause the game)

The pause menu resizes itself automatically and recenters itself

I readded the player's inventory. However, only the backpack works with interactions right now. Additionally the player inventory is not stored anywhere so it can't be transferred to the pause menu yet.